// packages/monomerDB/src/package-api.ts
/**
This file is auto-generated by the grok api command.
If you notice any changes, please push them to the repository.
Do not edit this file manually.
*/
import * as grok from 'datagrok-api/grok';
import * as DG from 'datagrok-api/dg';


export namespace queries {
  export async function getMonomerByGrokID(grokID: string ): Promise<DG.DataFrame> {
    return await grok.data.query('monomerDB:GetMonomerByGrokID', { grokID });
  }

  export async function upsertMonomer(symbol: string , name: string , monomerType: string , smiles: string , molblock: string , libraryName: string , username: string , polymerType: string | null, cappedSmiles: string | null, naturalAnalog: string | null, r1Id: number | null, r2Id: number | null, r3Id: number | null, r4Id: number | null, r5Id: number | null, r6Id: number | null, r7Id: number | null, r8Id: number | null): Promise<number> {
    return await grok.data.query('monomerDB:UpsertMonomer', { symbol, name, monomerType, smiles, molblock, libraryName, username, polymerType, cappedSmiles, naturalAnalog, r1Id, r2Id, r3Id, r4Id, r5Id, r6Id, r7Id, r8Id });
  }

  export async function createLibrary(name: string , username: string , friendlyName: string | null, description: string | null, libraryType: string | null, source: string | null, version: string | null): Promise<number> {
    return await grok.data.query('monomerDB:CreateLibrary', { name, username, friendlyName, description, libraryType, source, version });
  }

  export async function upsertMonomerProperty(monomerId: number , name: string , username: string , propertyType: string , valueString: string | null, valueNumber: number | null, valueBoolean: boolean | null): Promise<number> {
    return await grok.data.query('monomerDB:UpsertMonomerProperty', { monomerId, name, username, propertyType, valueString, valueNumber, valueBoolean });
  }

  export async function getMonomersByLibrary(libraryName: string ): Promise<DG.DataFrame> {
    return await grok.data.query('monomerDB:GetMonomersByLibrary', { libraryName });
  }

  export async function listLibraries(): Promise<DG.DataFrame> {
    return await grok.data.query('monomerDB:ListLibraries', {});
  }
}

export namespace funcs {
  export async function info(): Promise<void> {
    return await grok.functions.call('monomerDB:Info', {});
  }

  export async function getMonomerDBProvider(): Promise<any> {
    return await grok.functions.call('monomerDB:GetMonomerDBProvider', {});
  }

  export async function init(): Promise<void> {
    return await grok.functions.call('monomerDB:Init', {});
  }
}